Split disk and diskio source files
