+++ plugins/inputs/system/disk.go
@@ -2,8 +2,6 @@ package system
 
 import (
 	"fmt"
-	"log"
-	"regexp"
 	"strings"
 
 	"github.com/influxdata/telegraf"
@@ -82,144 +80,6 @@ func (s *DiskStats) Gather(acc telegraf.Accumulator) error {
 	return nil
 }
 
-type DiskIOStats struct {
-	ps PS
-
-	Devices          []string
-	DeviceTags       []string
-	NameTemplates    []string
-	SkipSerialNumber bool
-
-	infoCache map[string]diskInfoCache
-}
-
-func (_ *DiskIOStats) Description() string {
-	return "Read metrics about disk IO by device"
-}
-
-var diskIoSampleConfig = `
-  ## By default, telegraf will gather stats for all devices including
-  ## disk partitions.
-  ## Setting devices will restrict the stats to the specified devices.
-  # devices = ["sda", "sdb"]
-  ## Uncomment the following line if you need disk serial numbers.
-  # skip_serial_number = false
-  #
-  ## On systems which support it, device metadata can be added in the form of
-  ## tags.
-  ## Currently only Linux is supported via udev properties. You can view
-  ## available properties for a device by running:
-  ## 'udevadm info -q property -n /dev/sda'
-  # device_tags = ["ID_FS_TYPE", "ID_FS_USAGE"]
-  #
-  ## Using the same metadata source as device_tags, you can also customize the
-  ## name of the device via templates.
-  ## The 'name_templates' parameter is a list of templates to try and apply to
-  ## the device. The template may contain variables in the form of '$PROPERTY' or
-  ## '${PROPERTY}'. The first template which does not contain any variables not
-  ## present for the device is used as the device name tag.
-  ## The typical use case is for LVM volumes, to get the VG/LV name instead of
-  ## the near-meaningless DM-0 name.
-  # name_templates = ["$ID_FS_LABEL","$DM_VG_NAME/$DM_LV_NAME"]
-`
-
-func (_ *DiskIOStats) SampleConfig() string {
-	return diskIoSampleConfig
-}
-
-func (s *DiskIOStats) Gather(acc telegraf.Accumulator) error {
-	diskio, err := s.ps.DiskIO(s.Devices)
-	if err != nil {
-		return fmt.Errorf("error getting disk io info: %s", err)
-	}
-
-	for _, io := range diskio {
-		tags := map[string]string{}
-		tags["name"] = s.diskName(io.Name)
-		for t, v := range s.diskTags(io.Name) {
-			tags[t] = v
-		}
-		if !s.SkipSerialNumber {
-			if len(io.SerialNumber) != 0 {
-				tags["serial"] = io.SerialNumber
-			} else {
-				tags["serial"] = "unknown"
-			}
-		}
-
-		fields := map[string]interface{}{
-			"reads":            io.ReadCount,
-			"writes":           io.WriteCount,
-			"read_bytes":       io.ReadBytes,
-			"write_bytes":      io.WriteBytes,
-			"read_time":        io.ReadTime,
-			"write_time":       io.WriteTime,
-			"io_time":          io.IoTime,
-			"weighted_io_time": io.WeightedIO,
-			"iops_in_progress": io.IopsInProgress,
-		}
-		acc.AddCounter("diskio", fields, tags)
-	}
-
-	return nil
-}
-
-var varRegex = regexp.MustCompile(`\$(?:\w+|\{\w+\})`)
-
-func (s *DiskIOStats) diskName(devName string) string {
-	if len(s.NameTemplates) == 0 {
-		return devName
-	}
-
-	di, err := s.diskInfo(devName)
-	if err != nil {
-		log.Printf("W! Error gathering disk info: %s", err)
-		return devName
-	}
-
-	for _, nt := range s.NameTemplates {
-		miss := false
-		name := varRegex.ReplaceAllStringFunc(nt, func(sub string) string {
-			sub = sub[1:] // strip leading '$'
-			if sub[0] == '{' {
-				sub = sub[1 : len(sub)-1] // strip leading & trailing '{' '}'
-			}
-			if v, ok := di[sub]; ok {
-				return v
-			}
-			miss = true
-			return ""
-		})
-
-		if !miss {
-			return name
-		}
-	}
-
-	return devName
-}
-
-func (s *DiskIOStats) diskTags(devName string) map[string]string {
-	if len(s.DeviceTags) == 0 {
-		return nil
-	}
-
-	di, err := s.diskInfo(devName)
-	if err != nil {
-		log.Printf("W! Error gathering disk info: %s", err)
-		return nil
-	}
-
-	tags := map[string]string{}
-	for _, dt := range s.DeviceTags {
-		if v, ok := di[dt]; ok {
-			tags[dt] = v
-		}
-	}
-
-	return tags
-}
-
 type MountOptions []string
 
 func (opts MountOptions) Mode() string {
@@ -249,9 +109,5 @@ func init() {
 	ps := newSystemPS()
 	inputs.Add("disk", func() telegraf.Input {
 		return &DiskStats{ps: ps}
-	})
-
-	inputs.Add("diskio", func() telegraf.Input {
-		return &DiskIOStats{ps: ps, SkipSerialNumber: true}
 	})
 }
